feat: implement View bindings (#3)

+++ Models/SolutionOption.cs
@@ -3,5 +3,5 @@ namespace labcalcgit.Models;
 public class SolutionOption
 {
     public required string Name { get; set; }
-    public required double Percent { get; set; }
+    public required int Percent { get; set; }
 }
